fix(lib): make merge commit regex even less strict (fix #35)
From the git source [1] we can see that the only part that git generates for sure in the merge commit message is the
prefix `Merge `. This commit reduces the strictness of the merge detection to only that. While this might allow false
postives, I consider these better than false negatives (where committing is disallowed and people have to use
`--no-verify`).

[1]: https://github.com/git/git/blob/master/builtin/fmt-merge-msg.c#L418